Move classes used by code generators to codegen.*
