+++ gwtp-core/gwtp-dispatch-rest/src/main/java/com/gwtplatform/dispatch/rest/client/core/CoreModule.java
@@ -26,11 +26,6 @@ import com.gwtplatform.dispatch.rest.client.core.parameters.HttpParameterFactory
 public class CoreModule extends AbstractGinModule {
     @Override
     protected void configure() {
-        // TODO: Move all core classes to this package
-        // Classes used only by code-gen should go in core.codegen.*
-        // Root package should be for the public API
-        // Remove those silly `Rest` prefixes while in experimental
-
         bind(BodyFactory.class).to(DefaultBodyFactory.class).in(Singleton.class);
         bind(CookieManager.class).to(DefaultCookieManager.class).in(Singleton.class);
         bind(HeaderFactory.class).to(DefaultHeaderFactory.class).in(Singleton.class);
