+++ gwtp-core/gwtp-dispatch-rest/src/main/java/com/gwtplatform/dispatch/rest/client/core/DefaultUriFactory.java
@@ -22,9 +22,9 @@ import java.util.Map.Entry;
 
 import javax.inject.Inject;
 
+import com.gwtplatform.dispatch.rest.client.RestApplicationPath;
 import com.gwtplatform.dispatch.rest.client.annotations.GlobalQueryParams;
-import com.gwtplatform.dispatch.rest.client.annotations.RestApplicationPath;
-import com.gwtplatform.dispatch.rest.client.utils.RestParameterBindings;
+import com.gwtplatform.dispatch.rest.client.gin.RestParameterBindings;
 import com.gwtplatform.dispatch.rest.shared.HttpParameter;
 import com.gwtplatform.dispatch.rest.shared.HttpParameter.Type;
 import com.gwtplatform.dispatch.rest.shared.RestAction;
