+++ gwtp-core/gwtp-dispatch-rest/src/main/java/com/gwtplatform/dispatch/rest/client/interceptor/AbstractRestInterceptor.java
@@ -40,9 +40,9 @@ public abstract class AbstractRestInterceptor extends AbstractInterceptor<RestAc
     }
 
     @Override
-    public boolean canExecute(TypedAction action) {
+    public boolean canExecute(TypedAction<?> action) {
         if (action instanceof RestAction) {
-            InterceptorContext subjectContext = new InterceptorContext.Builder((RestAction) action).build();
+            InterceptorContext subjectContext = new InterceptorContext.Builder((RestAction<?>) action).build();
 
             for (InterceptorContext context : interceptorContexts) {
                 // Must have at least one supporting context
