+++ gwtp-core/gwtp-dispatch-rest/src/main/java/com/gwtplatform/dispatch/rest/rebind/entrypoint/DefaultEntryPointGenerator.java
@@ -25,8 +25,8 @@ import com.google.gwt.core.ext.UnableToCompleteException;
 import com.google.gwt.core.ext.typeinfo.JClassType;
 import com.google.gwt.user.rebind.ClassSourceFileComposerFactory;
 import com.google.gwt.user.rebind.SourceWriter;
-import com.gwtplatform.dispatch.rest.client.AbstractDispatchRestEntryPoint;
 import com.gwtplatform.dispatch.rest.client.DispatchRestEntryPoint;
+import com.gwtplatform.dispatch.rest.client.codegen.AbstractDispatchRestEntryPoint;
 import com.gwtplatform.dispatch.rest.rebind.AbstractGenerator;
 import com.gwtplatform.dispatch.rest.rebind.utils.ClassDefinition;
 import com.gwtplatform.dispatch.rest.rebind.utils.Logger;
