+++ gwtp-core/gwtp-dispatch-rest/src/main/resources/com/gwtplatform/dispatch/rest/rebind/subresource/SubResource.vm
@@ -1,6 +1,6 @@
 package $package;
 
-import com.gwtplatform.dispatch.rest.client.parameters.HttpParameterFactory;
+import com.gwtplatform.dispatch.rest.client.core.parameters.HttpParameterFactory;
 #foreach ($import in $imports)
 import $import;
 #end
