+++ gwtp-core/gwtp-dispatch-rest/src/test/java/com/gwtplatform/dispatch/rest/client/serialization/RestParameterBindingsSerializerTest.java
@@ -19,8 +19,9 @@ package com.gwtplatform.dispatch.rest.client.serialization;
 import org.junit.Test;
 
 import com.gwtplatform.dispatch.rest.client.core.parameters.HttpParameterFactory;
+import com.gwtplatform.dispatch.rest.client.gin.RestParameterBindings;
+import com.gwtplatform.dispatch.rest.client.gin.RestParameterBindingsSerializer;
 import com.gwtplatform.dispatch.rest.client.testutils.MockHttpParameterFactory;
-import com.gwtplatform.dispatch.rest.client.utils.RestParameterBindings;
 import com.gwtplatform.dispatch.rest.shared.HttpMethod;
 import com.gwtplatform.dispatch.rest.shared.HttpParameter.Type;
